Solution stored in "WFS150-7AR" folder is now capable of interfacing with SHA.
Although the exact problems that were occuring with the DLL are still unknown, it seems that the "Thorlabs.WFS.Interop64" DLL had a reference to a 32 bit DLL and was causing issues when VS would run the code for 64 bit architecture.  Switched to using "Thorlabs.WFS.Interop" instead and appeared to solve the problem.  Although I'm still not sure on this one as the error message was very, very vague.