Motor control now working.
I don't really understand the units that are going on with the API, but I was able to poll for some velocities and positions and got the motor to listen to me, and it appears to be going to the correct place (tested by moving manually to a location, homing, then using the script to return to that location).  Great Success!