Optimized code for automation, but created an error with the constructor.
Optimized the code for automation, which meant removing a lot of the user-specified values and replacing them with either hard-coded values or values coming from experiment_schedule.csv. In the process, though, there seems to be a problem with the code, in that it doesn't like creating two different constructors, but my understanding is that this is totally acceptable, so I'm wondering why that is a problem.