Initial Completion of ASEN_SHA, needs to be verified on Hardware.
I have completed the ASEN_SHA class, but it needs to be verified on the SHA hardware at SWRI before this can be confirmed. I simply went through the code and configured all the functions to run as we would like. This included setting the GatherCameraData function to take in the exposure time and output the image array, as well as the ProcessCameraData function to output the Zernikes. In addition, I configured the code to use the incoming beam width as the pupil size instead of defining it explicitly. Then, I ensured that the code wasn't cancelling out tip/tilt deviations, which it is set to cancel out by default. I also set the resolution of the WFS to be the maximum resolution supported by the WFS150-7AR, which is 1280x1024. Then, I set the reference distance from the MLA to the actual CCD to be the value given in the SHA manual. I removed any notion of highspeed mode, since this isn't supported by the WFS150-7AR. Finally, I made sure, in general, to limit the printing to the console in order to try and speed things up a bit.